refactor(core): split scale training into continuous, band, and color modules
Extract train-types, train-continuous, train-band, and train-color from the
386-line train.ts grab-bag. Facade preserves package-index and palette
identity re-exports (characterization in palettes tests). Co-locate
scales-m1 suites under tests/scales-m1/.

Fixes #552
