+++ .changeset/core-scales-train-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/core": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+refactor: split scale training into continuous, band, and color modules
